added span for .prompt stream
