+++ js/ai/src/prompt.ts
@@ -25,7 +25,7 @@ import {
   type JSONSchema7,
   type z,
 } from '@genkit-ai/core';
-import { Channel, lazy } from '@genkit-ai/core/async';
+import { lazy } from '@genkit-ai/core/async';
 import { logger } from '@genkit-ai/core/logging';
 import type { Registry } from '@genkit-ai/core/registry';
 import { toJsonSchema } from '@genkit-ai/core/schema';
@@ -35,7 +35,6 @@ import { existsSync, readFileSync, readdirSync } from 'fs';
 import { basename, join, resolve } from 'path';
 import type { DocumentData } from './document.js';
 import {
-  GenerateResponseChunk,
   generate,
   generateStream,
   toGenerateActionOptions,
@@ -449,8 +448,6 @@ function wrapInExecutablePrompt<
     input?: I,
     opts?: PromptGenerateOptions<O, CustomOptions>
   ): Promise<GenerateResponse<z.infer<O>>> => {
-    logger.debug('EXEC');
-
     return await runInNewSpan(
       registry,
       {
@@ -476,8 +473,6 @@ function wrapInExecutablePrompt<
     input?: I,
     opts?: PromptGenerateOptions<O, CustomOptions>
   ): Promise<GenerateOptions<O, CustomOptions>> => {
-    logger.debug('\n\nRENDER\n\n');
-
     return {
       ...(await renderOptionsFn(input, opts)),
     } as GenerateOptions<O, CustomOptions>;
@@ -487,49 +482,7 @@ function wrapInExecutablePrompt<
     input?: I,
     opts?: PromptGenerateOptions<O, CustomOptions>
   ): GenerateStreamResponse<z.infer<O>> => {
-    let channel = new Channel<GenerateResponseChunk>();
-
-    const streamed = rendererAction.then(async (promptAction) => {
-      const generated = runInNewSpan(
-        registry,
-        {
-          metadata: {
-            name: promptAction.__action.name,
-            input,
-          },
-          labels: {
-            [SPAN_TYPE_ATTR]: 'util',
-          },
-        },
-        async (metadata) => {
-          console.log('\n\nABOUT TO GENERATE DA STREAM\n\n');
-
-          const { response: res, stream: str } = generateStream(
-            registry,
-            renderOptionsFn(input, opts)
-          );
-          for await (let chunk of str) {
-            console.log(`stream chunk: ${chunk}`);
-            logger.info(chunk);
-            channel.send(chunk);
-          }
-          //metadata.output = await response;
-          return res;
-        }
-      ).then((res) => res);
-
-      generated.then(
-        () => channel.close(),
-        (err) => channel.error(err)
-      );
-
-      return generated;
-    });
-
-    return {
-      response: streamed,
-      stream: channel,
-    };
+    return generateStream(registry, renderOptionsFn(input, opts));
   };
 
   executablePrompt.asTool = async (): Promise<ToolAction<I, O>> => {
