--- js/testapps/dev-ui-gallery/prompts/code.prompt
@@ -1,21 +1,0 @@
----
-model: googleai/gemini-2.0-flash
-config:
-  temperature: 0.4
-  safetySettings:
-    - category: HARM_CATEGORY_HATE_SPEECH
-      threshold: BLOCK_ONLY_HIGH
-    - category: HARM_CATEGORY_DANGEROUS_CONTENT
-      threshold: BLOCK_ONLY_HIGH
-    - category: HARM_CATEGORY_HARASSMENT
-      threshold: BLOCK_ONLY_HIGH
-    - category: HARM_CATEGORY_SEXUALLY_EXPLICIT
-      threshold: BLOCK_ONLY_HIGH
-input:
-  schema:
-    language: string
-  default:
-    persona: java
----
-
-Generate a program that calculates fibonacci in {{language}}.
